requirements.txt
openai-whisper
ffmpeg-python
opencv-python
pillow
